Light Telemetry and redesigned Settings Window
I added slight Telemetry features that send anonymous data.

The Data send contains a UUID (Universally Unique Identifier) which gets generated on first startup. 
Current time, Software Version and if the software is running or not.

No service running in the background, it just sends a ping on startup and another one on shutdown. 

You can always disable telemetry in the settings window which also has been a bit redesigned, still intuitive though :)