Spotify API - Split Songinfo / Albumcover
You can now save artist and title in different text files, but still maintain the combined text file. 

Also when using the Spotify API you can now enable that the album cover will be downloaded. The cover is in the same directory as the text file and is named "cover.jpg"