Fixed an issue that crashed the program when no Twitch credentials were present and it tried to connect to the twitch services.
Changed the SpotifyAPI connection to be manually triggered instead of autmatically when starting the software.

Added a tooltip to the little button on the twitch connection settings that opens an external site for generating the oAuth token

Added error logging to the history

Changed the layout on the settings window a bit, nothing major.