Additions: Added Queue Window - here you can see the current song queue - you can delete single requests (note: you still have to remove them from the spotify queue)
Added Blacklist Window
- here you can set blacklisted Artist (these won't get added to your queue on request)

Added a WebHelper Class to reduce using the same code as hardcode

Fixes:
Fixed a potential crash when writing to the log file.

Fixed a crash when connecting to twitch but the connection was already established.

Fixed songs not being saved to the web-queue properly

Changes:
Changed the way the software stores its positon. It now opens where you closed it.
Changed the way how sub-windows open. They open at the same location as the main window now.
Changed the way the Twitch Connection settings are getting saved. Now you don't have to click a button that had no response.