You can now add the requester to the output. For that add {{custom text{req}}}. The text inside {{ and }} will only be displayed if it is a requested song.
Changed the way the pause text works when using the spotify API. The API is not 100% consistent and sometimes gives out that there is nothing is playing even though it is. The pause text will now appear 10 seconds after you paused. This way I can make sure thats not a false positive by spotify's api.